\name{twoloops}
\alias{twoloops}
\docType{data}
\title{
Consanguineous pedigree example
}
\description{
A consagnuineous pedigree.
}
\usage{data(twoloops)}
\format{
A data frame describing the pedigree in standard LINKAGE format.
}
\examples{
data(twoloops)
x = linkdat(twoloops)
plot(x)
}
\keyword{datasets}
